add authorization for game patch and delete.  User must be game creator

+++ 4_client_server_db_jwt/server/resources/game.py
@@ -1,10 +1,13 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import (
+    get_jwt_identity,
+    jwt_required,
+)
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
-from models import Round, Game
+from models import Round, Game, User
 from schemas import GameSchema, GameUpdateSchema, RoundSchema
 
 blp = Blueprint("Games", "games", description="Operations on games")
@@ -22,9 +25,12 @@ class Games(MethodView):
     @blp.arguments(GameSchema)
     @blp.response(201, GameSchema)
     def post(self, fields):
-        """Create a new game and 1st round of play"""
+        """Create a new game and 1st round of play for the current user"""
         try:
+            user_id = get_jwt_identity()
+            user = db.get_or_404(User, user_id)
             game = Game(**fields)
+            game.user = user
             db.session.add(game)
             db.session.commit()
             round = Round(game = game, range_min = game.range_min, range_max = game.range_max)
@@ -49,8 +55,11 @@ class GamesById(MethodView):
     @blp.arguments(GameUpdateSchema)
     @blp.response(200, GameSchema)
     def patch(self, fields, game_id):
-        """Update game by id.  Update current round based on the guess."""
+        """Update game by id.  Update current round based on the guess. Authorization requires user to be game creator. """
         game = db.get_or_404(Game, game_id)
+        user_id = get_jwt_identity()
+        if game.user_id != user_id:
+            abort(403, message=f"You do not have permission to modify Game {game_id}.") 
         if game.is_over:
             abort(409, message=f"Game {game_id} is over.")
         try:
@@ -68,8 +77,11 @@ class GamesById(MethodView):
     @blp.doc(authorize=True)
     @blp.response(204)
     def delete(self, game_id):
-        """Delete game and associated rounds by id"""
+        """Delete game and associated rounds by id.  Authorization requires user to be game creator."""
         game = db.get_or_404(Game, game_id)
+        user_id = get_jwt_identity()
+        if game.user_id != user_id:
+            abort(403, message=f"You do not have permission to delete Game {game_id}.")
         try:
             db.session.delete(game)
             db.session.commit()
